move manipulator into agents
