+++ habitat-lab/habitat/articulated_agents/robots/static_manipulator.py
@@ -8,7 +8,7 @@ import attr
 import magnum as mn
 import numpy as np
 
-from habitat.articulated_agents.robots.manipulator import Manipulator
+from habitat.articulated_agents.manipulator import Manipulator
 from habitat_sim.simulator import Simulator
